+++ node/resolvers/search/index.ts
@@ -435,6 +435,10 @@ export const queries = {
 
     const result = await biggySearch.facets(biggyArgs)
 
+    if (ctx.vtex.tenant) {
+      ctx.vtex.tenant.locale = result.locale
+    }
+
     // FIXME: This is used to sort values based on catalog API.
     // Remove it when it is not necessary anymore
     if (result && result.attributes) {
@@ -566,6 +570,10 @@ export const queries = {
 
     const products = await biggySearch.productSearch(biggyArgs)
 
+    if (ctx.vtex.tenant) {
+      ctx.vtex.tenant.locale = products.locale
+    }
+
     const regionId = segment?.regionId
     const convertedProducts = await productsBiggy({ ctx, simulationBehavior, searchResult: products, regionId })
     convertedProducts.forEach(product => product.cacheId = `sae-productSearch-${product.cacheId || product.linkText}`)
@@ -660,6 +668,10 @@ export const queries = {
     }
 
     const result = await biggySearch.productSearch(biggyArgs)
+
+    if (ctx.vtex.tenant) {
+      ctx.vtex.tenant.locale = result.locale
+    }
 
     const productResolver = args.productOriginVtex
       ? productsCatalog
@@ -775,6 +787,10 @@ export const queries = {
       salesChannel: tradePolicy,
       sellers
     })
+
+    if (ctx.vtex.tenant) {
+      ctx.vtex.tenant.locale = result.locale
+    }
 
     const productResolver = args.productOriginVtex
       ? productsCatalog