+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Use the locale returned by the search queries.
+
 ## [1.39.2] - 2021-05-17
 
 ### Fixed